energy/pi_zero_energy/main.py
#!/usr/bin/env python3
"""
Energy Monitor - Battery management and monitoring
"""

import sys
from pathlib import Path

sys.path.insert(0, str(Path(__file__).parent.parent.parent / 'shared'))

import yaml
from logger import setup_logger

logger = setup_logger(__name__, 'logs/energy.log')

def load_config():
    config_path = Path(__file__).parent / 'config.yaml'
    with open(config_path, 'r') as f:
        return yaml.safe_load(f)

def main():
    logger.info("Energy Monitor - Starting")
    
    config = load_config()
    
    # TODO: Initialize
    # - INA219 sensor
    # - MQTT client
    
    logger.info("Monitoring started")
    
    try:
        import time
        while True:
            # TODO: Read battery metrics
            # - Voltage
            # - Current
            # - Power
            # - Calculate %
            
            time.sleep(config['monitoring']['interval'])
    except KeyboardInterrupt:
        logger.info("Shutdown")

if __name__ == '__main__':
    main()
